+++ src/examples/simple_zkapp.js
@@ -51,7 +51,7 @@ let initialState = Field(1);
 let zkapp = new SimpleZkapp(zkappAddress);
 
 console.log('compile');
-await SimpleZkapp.compile(zkappAddress);
+await SimpleZkapp.compile();
 
 console.log('deploy');
 let tx = await Mina.transaction(feePayerKey, () => {